llm_evaluation/requirements.txt
absl-py==2.1.0
accelerate==0.26.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
allennlp==1.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.5.2
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
benepar==0.2.0
bert-score==0.3.9
bitsandbytes==0.41.3
black==21.7b0
blanc==0.3.4
blis==0.4.1
boto3==1.35.57
botocore==1.35.57
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==7.1.2
cmake==3.30.4
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.6
Cython==0.29.32
dataclasses==0.6
datasets==1.7.0
debugpy==1.8.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker==5.0.0
edlib==1.3.9.post1
emoji==2.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl#sha256=dfba97b3cb5c177f60c4087a5b4f5c9d7f5c0fce7150d8aa7c0aab4eb4811a07
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fairseq==0.10.2
filelock==3.0.12
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gin==0.1.6
gin-config==0.5.0
googledrivedownloader==0.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonnet==0.20.0
jsonpickle==4.0.0
jsons==1.6.3
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lit==18.1.8
lxml==5.3.0
marisa-trie==1.2.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
moverscore==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.8
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.53.0
openpyxl==3.1.5
overrides==3.1.0
packaging==24.1
pandas==1.3.3
parameterized==0.8.1
parso==0.8.4
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkg_resources==0.0.0
plac==1.1.3
platformdirs==4.3.6
pluggy==0.13.1
portalocker==2.10.1
preshed==3.0.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyahocorasick==2.1.0
pyarrow==3.0.0
pycparser==2.22
pydantic==1.7.4
pydantic_core==2.23.4
pyemd==0.5.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==6.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-pretrained-bert==0.6.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qa_metrics==0.2.23
questeval==0.2.4
regex==2024.9.11
repro==0.1.1
requests==2.32.3
rouge-score==0.1.2
s3transfer==0.10.3
sacrebleu==2.0.0
sacremoses==0.1.1
sacrerouge==0.2.5
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.2.1
sentencepiece==0.1.95
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.0.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
stanza==1.9.2
summ-eval==0.892
summac==0.0.4
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
textsearch==0.0.24
thinc==8.0.17
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
tomli==1.2.3
torch==2.4.1
torch-struct==0.5
torchvision==0.19.1
tornado==6.4.1
tqdm==4.49.0
traitlets==5.14.3
transformers==4.46.3
triton==3.0.0
typer==0.3.2
typing==3.7.4.3
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.2
Unidecode==1.2.0
urllib3==1.26.20
wasabi==0.10.1
wcwidth==0.2.13
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.14.0
zipp==3.20.2
